[GAIA-1358] changed byte conversion (dynabuffers-python), added signed support

+++ dynabuffers-python/dynabuffers/ast/datatype/LongType.py
@@ -7,7 +7,7 @@ class LongType(ISerializable):
         return 8
 
     def serialize(self, value, buffer, registry):
-        buffer.putFloat(value)
+        buffer.putLong(value)
 
     def deserialize(self, buffer, registry):
-        return buffer.getFloat()
+        return buffer.getLong()
